Fix bug 175 - Grapical EPG, Change timer, Delete hot key is green should be red.
**Requires Fix-bug-175 (commit 226a589) in easy-ui-4**

References new screen RecordTimerQuestion1 in easy-skin-aus-hd/skin.xml
in easy-ui-4.

Changed the buttons on the Change Timer menu to be RED Delete timer
and GREEN Edit timer. Also made the entry for YELLOW Disable Timer
visible. The menu on the REC button remains GREEN Add Timer and
BLUE Add AutoTimer.

The change in visible length of the Change Timer menu required the
addition of screen RecordTimerQuestion1 to skin.xml.

The menus are now longer, and a cleanup of the menu location code
will follow in another commit.