Tidy up REC and GREEN Change timer menus in EPG
**Requires Fix-bug-175 (commit 226a589) in easy-ui-4**

The addition of a third entry into the GREEN Change timer menu meant
that the menu now covered more of the bottom button hint bar,
especially in the InfoBar EPG.

Changed the display of the menu so that if there is room it is a
dropdown menu below the EPG entry and no longer obscures part of
the EPG entry for the event which the timer is being set. If there
isn't enough room below the event entry for the menu, the menu
appears on the top edge of the EPG entry.