Fix bug #248  -  Change Timer menu not removed in Event View
Added EventViewEPGSelect.showChoiceBoxDialog() and
EventViewEPGSelect.closeChoiceBoxDialog() to correctly enable and
disable ActionMaps added in the child class.  The methods call their
corresponding parent class methods in EventViewBase to complete the
action.

This prevents the binding of the GREEN key in EventViewEPGSelect
from affecting the operation of the Change Timer menu launched from
EventViewBase.

This in turn prevents the second Change Timer menu from being created
on the press of GREEN in the Change Timer menu. This second menu
takes the place of the first one in the EventViewBase instance,
leaving the first menu on the screen. The GREEN key binding for the
orphaned menu cause a crash when it is invoked by another press of
GREEN.