Improve "stop recording" facilities
In the previous code REC or RED was pressed in live TV and the menu
item Stop timer recording was pressed, or TIMER was pressed in live
TV,  the Timer Edit screen was entered.

This screen allowed repeated timer recordings to be stopped (and
disabled) by navigating to the timer entry and pressing Disable,
which gave a menu that allowed the recording to be stopped or
disabled.

The screen did not allow non-repeated recordings to be stopped.

This commit changes the YELLOW button hint text to "Stop/Disable"
when a repeated timer is the focus, and to "Stop" when a non-repeated
timer is the focus. "Stop/Disable" behaves in the same way that
"Disable" did for a repeated timer. "Stop" simply stops the recording
in focus.