AudioSelection: Also enable volumecontrol when using B/CH/P +/- buttons.
When the normal volume buttons are re-assigned to the TV (via HDMI_CEC or a
programmable control) there is no way to control the volume. In the
AudioSelection the B+/-, CH+/- or P+/- buttons are not used. Now use these
buttons to control the box volume when we are in AudioSelection menu. Also
enable mute button in AudioSelection.