Enhancement 268 - Add drive information to MENU>Setup>Information>Devices
Added HDD drive information to the volume information in the Devices
Information screen.

Removed unused code that performed the same function from the About
screen.

Added file system type information to volumes.

Rationalised the placement of newlines in the generated text.

Simplified code that reads partition table.

Removed code that reads the mount table, and replaced it wth an
existing function.

Returned to using 'mountpoint' as a list of strings instead of a
single string.

Use existing function to format HDD sizes instead of inline code.