Simplify ServiceInfo class and make ServiceInfoSimple display consistent with ServiceInfo Multiplex display
Screens/ServiceInfo.py:

Combine the better aspects of the techniques used for display list generation for the ServiceInfoSimple Live TV display with those used for ServiceInfo Multiplex and Tuner displays.

This has the side-effect of giving them consistent labels and ordering.

The list-based generation meant that the generation of a tuner name and the channel number were pushed down into Tools.Transponder.ConvertToHumanReadable().

Tools/Transponder.py:

Generate the tuner name and the channel number in ConvertToHumanReadable()