Fix bug #282  -  New mail notification in GMailReader plugin grashes GUI
Crashed when a file to record the last message count could not be
read. The file could only be created and written to for the first
time after it had been read. The same function was expected to
return a value but did not

Replaced the use of a file to store the last message count with a
config variable. Added a return to the function.