Enhancement 283 - Add boot loader version information to MENU>Setup>Information>About
Use "strings | grep" to extract the OEM build date for the boot
loader from the boot loader binary and display it as a Unix ctime
time/date string. The more recent of the two datestamps is used,
the other one is assumed to be the CFE timestamp. Both timestamps
are available to the calling code, but only the OEM date is displayed.

Code to display a short form for the boot build date/time has also
been written as an alternative, and is commented out.