Also load userbouquets that are not in bouquets.tv or bouquets.radio
This allows to install downloadable settings but keep your own favorites and include these automatically into the downloaded settings.
This is done by read the userbouquets files from the settings and parse the standard root file (bouquets.tv and bouquets.radio).
While parcing the root files already readed bouquets will be removed from the collected list and afterwards the bouquets that
are left over are loaded and pushed in front of the settings. I move them in front as I suggest these bouquets should be most
likely prefered bouquets.

In addition I removed the channelnumbering from the userbouquets parsing routine and call remumber instead after the userbouquets
are loaded.

In addition when a user bouquet is removed also the file will be renamed by adding the extension ¨.del". This to avoid that the
next time when enigma2 is restarted removed bouquets are reloaded. By manually removing the ".del" extension these bouquets
could be restored after a next reload settings or restart of enigma2

The added userbouquets must have references in the lamedb of course to get this functioning perfectly.

It might be that a lot of users can get extra bouquets in their settings. When these added bouquets are not wished the user can
manually remove them once and afterward delete al the ".del" files from the /etc/enigma2 directory.