Fix #168 - First navigation in Graphical EPG leaves wrong fill in event
The problem was caused by clip rectangles set by
Components.EpgList.selEntry() being set incorrectly when the EPG
was being set up, because Components.EpgList.setShowServiceMode()
was called to late (after Components.EpgList.fillGraphEPG()) in
Screens.EpgSelection.onCreate().

The calls of setShowServiceMode() were moved to before the call of
fillGraphEPG().

Another bug, initially thought to be related to this problem, was
also fixed.  In Components.EpgList.selEntry(), the height of
self.select_rect was set to self.event_rect.height (reference to a
function returning the height instead of self.event_rect.h (the
height).