Fix bug #288  -  Incorrect left fill in current service row in graphical EPG
Bug 288 is caused by the fact that there's a pixmap underlying the
whole of each row and it's supposed to fill in for missing bits of
the EPG. The problem is that two different pixmaps are used.
epg/OtherEvent.png (navy blue) is used for all rows except the
currently showing live TV service. epg/SelectedEvent.png (bright
yellow) is used for the current service. The yellow of
epg/SelectedEvent.png shines through wherever there are gaps in the
EPG for the current service.

The left edge of the EPG is where this usually occurs.

There is no corresponding bug in the text mode of the graphical
EPG.

The fix is to only use epg/OtherEvent.png as the background fill
for graphical EPG rows.