Fix bug #168 - First navigation in Graphical EPG leaves wrong fill in event - second try
The issue was that in Components.EPGList, findBestEvent(), which is
used to find where to place the selection when you move between
services, was getting called by fillGraphEPG() in the EPG startup
code before the current service had been set by moveToService().
This meant that findBestEvent() would initially search for an event
in the event list of the first entry in the services list. This
would be corrected by later calls to findBestEvent(), and the
selection highlight would be displayed correctly. However,
moveToService() calls selEntry(), which in turn calls
eListboxPythonMultiContent.setSelectionClip(). The incorrect
selections in findBestEvent() would then set up an incorrect set
of clipping rectangles for the selection highlight, so that when
the display was selectively redrawn on a press of a navigation
button, some of the highlight for the old selection wasn't over-written.

Unfortunately, other code in Components.EPGList depends on there
being a call of findBestEvent() in fillGraphEPG(), so simply removing
it wasn't an option.

Instead, a flag variable has been added to Components.EPGList that
prevents findBestEvent() and selEntry() from acting if the current
service hasn't been initialised by moveToService(), and A to call
findBestEvent() has been added to moveToService().

It's a bit of an inelegant solution, but there didn't appear to be
any state information in the class that could be used to check
whether the current service had been initialised. The variable
cur_service cannot be used for this purpose, and in fact the way
it is set in findBestEvent() is part of the cause of the problem.

Fixing the bug this way also independently fixes the most visible
apparent cause of the problem, the one addressed in commit a146dd2.
However, that was only a partial fix of bug 168. Because of this,
this commit reverses the changes made to Screens/EpgSelection.py
in a146dd2.

Also fixed a minor formatting typo in selEntry, in te construction
of self.select_rect.

All unqualified methods above are methods of Components.EPGList.

Another bug in highlighting remains. It is intermittent but can
often be triggered by 8 (page down), then DOWN in the graphical
EPG. It seems only related to this bug by appearing to be another
problem with the clipping state in eListboxPythonMultiContent. I
have been unable to trace it to errors in the Python code.