Introduce numberzap functionality in channelselection screen
This means the hotkeys numbers cannot be used anymore as the numberkeys
are needed for the numberzap. But in the contextmenu we also have these
hotkeys recently. Also replaced focus to currently played service to the
contextmenu. Did not add the focus to latest service in history. I suggest
this was somehow useless. In addition I added a confirmation before
service delete as it was on the hotkey.