Implement feature request #89 - Setup wizard offer to format hard disk - improve error handling
Add callbacks from format jobs to gather status information sbout
the formatting job.

If there is an error message, return to the format setupscreen with
the text box changed to idicate that there has been an error and a
panel that displayes any error messages.

Requires matching commit 4087280 in branch implement-feature-89 in
skin-default repository.