Fix bug 199 - Timeshift Save Recording (Stop after current event) fails when channel changes
Handle case when "Show warning when timeshift is stopped" is disabled,
the user asks record current event from timeshift, and then changes
channel.

Also fixed the ancilliary bug that causes the "referenced before
assignment" when the file in the first part of the error message
can't be found. That fix should not be accessible from the action
sequence that caused the main bug now.