Fix bug #308 - Problems with setting default EPG type when the Movie Player is used
[InfoBarGenerics, UsageConfig]
Move creation of config variables config.usage.defaultEPGType and
config.usage.defaultGuideType out of InfoBarEPG and into
UsageConfig.InitUsageConfig() so that the variables are only
initialised once.

Some changes were needed to ensure that the saved value of the
variable was correctly carreied through from the creation of the
variables to their use.

[InfoBar, InfoBarGenerics, UsageConfig]
Try to clarify the naming of the config variables (defaultEPGType
sets the guide style for the INFO button and defaultGuideType sets
the guide style for the EPG button) by using clearer local variable
names for the config variables.