Introduce fall-back using remote enigma2 receiver.
First shot at a feature to enable using another enigma2 receiver
as fall-back. The idea is that enigma2 uses a tuner from another
receiver automatically, whenever a service cannot be selected
(view/record/stream/pip) because no tuner is available (either
because it is busy with another service or because there are no
tuners that can handle this service (dvb-s/dvb-c/dvb-t) etc.)

At least for the moment this has one major drawback: when a
service cannot be selected locally, the remote service is always
started, even if the remote receiver cannot handle the service.
Unfortunately there is, afaik, no mechanism to probe a remote
enigma2 instance whether a service "is playable" like there is for
local services.

If no "remote fallback tuner" is configured (default), service
selection works as usual and the above drawback does not apply.