Fix bug #320 - REC, "Timeshift save recording (stop after current event)" doesn't always save timeshift on zap
If the user has asked to save the timeshifted event with REC,
"Timeshift save recording (stop after current event)" and
MENU>Setup>TV>Timeshift settings>Timeshift-save action on zap is
set to anything other than "Ask user", and
MENU>Setup>TV>Timeshift settings>Show warning when timeshift is stopped
is set to On, perform a savetimeshiftandrecord action on zapping so
that the timeshift event is recorded as the user asked.

This is consistent with the behaviour when Timeshift-save action on zap
is set to anything other than "Ask user" and
Show warning when timeshift is stopped
is set to Off.