Fix bug #316 - Previous Channel button exits Timeshift - add help info
[InfoBarGenerics] Add help info to NumberActions NumberActionMap.
The InfoBarNumberZap.keyNumberGlobal() has states where there is
no action.  In these states, the method returns 0, so that if the
keypress is not handled by some lower priority ActionMap, the
UnhandledKey popup is shown.  In those states, the help generation
method InfoBarNumberZap.helpKeyNumberGlobal() returns None to
indicate that the key has no function in the current state.

[HelpMenuList] Ignore any help entry with value None.