Fix Bug #328 - Incorrect action mapping or contexts for INFO button in some screens.
[UpdatePluginMenu] Removed InfobarEPGActions context from
self["shortcuts"].  It was unused and causing unnecessary warnings
on INFO/break.

[PluginManager] Remove unused InfobarEPGActions context from
self["shortcuts"].  Changed the action name for INFO to WizardActions/info
so that it becomes active. Split the packageInfo() method out of
handleSelected() so that INFO only acts to show information in the
packages lists and doesn't activate Update in the categories list.
Added help text to indicate when INFO is active.

[EventViewBase] Remove insaccessible "info" action from self["actions"].
INFO already acts through EventViewActions/contextMenu. This change
does not affect class behaviour.