Fix Bug #332  -  No TEXT button hint in MiniDLNA setup screen
Cleaned up a lot of the code around the dynamic display of the TEXT
button hint.

Modified ConfigListScreen so that it handles self["VKeyIcon"] as
either a Boolean (as it did before) or as a Pixmap (which was
previously handled by overriding ConfigListScreen.handleInputHelpers()).

Removed now-redundant code from several classes in [NetworkSetup]
and [Setup] to standardise handling of the TEXT button hint.

Added self["VKeyIcon"] as Pixmap (or changed it from Boolean to
Pixmap) in Setup, NetworkSambaSetup, NetworkuShareSetup and
NetworkMiniDLNASetup so that the TEXT button hint is displayed
through the standard mechanism.

This simplified the TEXT button hint display and made it more
uniform, as well as fixing bug 332.

Requires matching checkin 86ae7e8 in branch Fix-bug-332 in skin-default
repository.