Fix Bug #334  -  Recurring exceptions and other bugs in [IconCheckPoller]
Use already existing methods to enumerate the network interfaces and
get their status.

Refactor both the Network and USB state update to execute nothing if the
symbol files don't exist, and avoid enumerating and checking devices
if the result won't be used. Also break out of the check loop as soon
as the status is set to 1.