Fix Bug #301 - Allow HelpableActionMaps to have multiple contexts
[ActionMap] Allow HelpableActionMap and HelpableNumberActionMap
to accept a list of contexts as well as a single context. This
makes it possible to convert ActionMap and NumberActionMap
objects to helpable without changing their basic structure.

Single-context helpable ActionMaps are retained for backwards
compatibility.

[HelpMenuList] Change help grouping indexing from context to
id(actionmap). This was necessary to handle multi-context helpable
ActionMaps. It is how the implementation should have been in the first place.

[IniLastFM] Change self["actions"] LastFMScreenMain to be a multi-context
HelpableActionMap and remove the workaround that had previously been
used to add help to a multi-context ActionMap when only single-context
HelpableActionMaps were supported.

Also changed self["actions"] in LastFMSaveScreen to a HelpableActionMap
and added help.

These changes are for a demonstration of the new capability and
for testing. Full use to add help across the UI is a larger job.