Fix Bug #301 - Automate flagging of action help for configurable actions
[HelpMenuList] Use the existance of a function to return help info
for an action to flag the action as configurable.

[HelpMenu] Mention the configurability flagging in the HELP screen
for help.

[EpgSelection, InfoBar, InfoBarGenerics, MovieSelection] Remove old
manual flagging of actions as configurable.

Use get() in help lookups instead of directly indexing with [] to
give soft failure if there's a missing key in the help dict.

Changed "Skip backward" to "Skip back" to make text for skips more
uniform.