Fix Bug #168 - First navigation in Graphical EPG leaves wrong fill in event
[EpgSelection] in Screens.EPGSelection.onCreate(),
Components.EPGList.selEntry() was first called (via EPGList.moveToService()
and EPGList.findBestEvent()) before the correct configured values
of EPGList.showServiceTitle and EPGList.showPicon were set (in
EPGList.setShowServiceMode()).

This meant that when EPGList.selEntry() was first called, and it
in turn called EPGList.recalcEntrySize(), recalcEntrySize() used
possibly incorrect values of EPGList.showServiceTitle and
EPGList.showPicon to calculate EPGList.service_rect and EPGList.event_rect.

Then, the first call of selEntry() could set the selection clip
rectangle at the wrong location, and that lead to incorrect fill
being left after the first navigation.

Moved the calls of EPGList.setShowServiceMode() to before the call
of EPGList.moveToService() so that when EPGList.selEntry() is first
called it uses the correct location for EPGList.event_rect when it
calculates the selection clipping rectangle.