[Subtitles] PangaSubtitles: Arrange that subtitles can be displayed Italic or Bold
And herefore remove alternative colloring as previously Italic was displayed as yellow
Still it does not feature everything that is possible in srt files as only whole lines
could be displayed in Italic and Bold and the font switch could not be made on a segment
of a line, but most subtitles do not have that.
Thanks PLi!