Fix Bug #351 - You Tube main screen button labelling incorrect and other UI problems
[MyTubePlayerMainScreen]
Move initialisation of text in buttons to LayoutFinished time,
so that later hide()/show() operations on the buttons work correctly.

Change the skin and other definitions of the BLUE button to be the same as
other buttons/button hint.

Make hide()/show() operations on the buttons consistent with the actual
enabling/disabling of the screen's aActionMaps.

Requires supporting skin changes in commit 7e973ac in branch Fix-bug-351
in skin-default repository.