Fix Bug #351 - You Tube main screen button labelling incorrect and other UI problems
[MyTubePlayerMainScreen]
Normalised text input to search box and enabled correct display
of TEXT button hint and SMS-style input help window. The old code allowed
SMS-style input to the search box only when it was enabled (CH+ from
browsing feedlist), but the TEXT button was active to open the
on-screen virtual keyboard in all contexts except the search history list.

The code that was intended to add opening the on-screen virtual keyboard
didn't actually work, the opening of the on-screen keyboard fell back to
the ActionMaps in [ConfigListScreen].

Made text input only active when the search box is active, and display
the TEXT button hint and the SMS hint window when text input is active.

[MyTubeService.GoogleSuggestions]
The code made use of an apperrently outdated way to enabling debugging
in HTTPConnection. This caused Google suggestions to fail, so that
no suggestions were presented in [MyTubePlayerMainScreen]. Changed the
debug level setting to reflect the current API for HTTPConnection.

Requires supporting skin changes in commit 7e973ac in branch Fix-bug-351
in skin-default repository.