base/filepush.cpp: Stop background thread in eFilePushThread destructor
The eThread class has a design fault: Classes are supposed to inherit
from it. This can't work in C++, because once the eThread d'tor fires,
the object has alread been demoted and partially destroyed. This
results in the thread now running with halfway destroyed data.

A workaround is to always join the thread in each class that inherits
from eThread, which is what this commit implements.

A proper fix would be to remove all virtual methods from eThread, and
use a "runnable" or similarly named member to actually implement the
thread's code.