filepush: Get rid of eThread::sync
The sync call is way to complicated for what it is used for here. All the
client wants to know is "did I start yet?" which it can easier tell by
looking at its own "m_stop" value.

The intention here is to get rid of the eThread::sync method alltogether.