epgcache: Remove eTread::sync usage
All it wants to know is whether it can start sending messages to the
thread, which is already the case even before the thread runs. Just
use the m_running member for this, hence also fixing the fact that this
boolean was written in another thread without a memory barrier.