Fixed: adjusting of external subtitles delay required video restart
Bug description: changing of delay for external subtitles required a
restart of video. This was because the delay setting (and the fps
conversion as well) was applied during reading subtitles from disk,
which happen when the playback starts.
Fix: the delay setting is now taken into account when the subtitles are
pushed to screen. This allows for instant effect.
Bonus fix: the delay setting read from config file must be divided by
90 because it is stored in pts, not in ms.