epgcache.cpp: Fix possible lock scope error and hide channel_map_lock
No need to export channel_map_lock to the outside world, hide it in a static.

The channel_map_lock apparently protects m_knownChannels, but why it sometimes
locks and sometimes doesn't is a riddle. In DVBChannelStateChanged it looks
weird, as it aquires a lock but doesn't release it until after unlocking the
"channel_active". Looks as if a scope is missing here, add it.