Fix Bug #356 - [Pixmap] KodiDirect plugin crashes after startup splash screen(s)
The KodiDirect plugin crashes after showing its startup splash
screens.

The underlying cause is that KodiDirect starts making method calls
on the instance objects in the MovingPixmaps in self["frame0"] and
self["frame0"] before the instance objects are created.

Earlier code in MovingPixmaps protected the code from crashing if
the instance object was invalid. That protection was removed in
commit 2588a321, and the protection has been reinstated in this commit.