Front panel tweaks
In LCD display skins, numerous tweaks to prevent overlap of items and
ensure that all screen items fall within the screen area.

[MovieSelection] Fix bug in MovieContextMenuSummary where selectionChanged
indexed the selected item incorrectly and so the intended text was
not displayed. Also changed the title of the Movie Player MENU screen
to differentiate it from the MENU>Settings... screen. Also removed
redundant locals translation calls from the titles of teh two screens.