Revert "ChannelSelection: Make it possible to hide userbouquets"
This reverts commit 323f89f3487a9f9a8c8f12e840c6c47825183286.

As this patch was not really required. In OpenPLi hide bouquets was functioning
directly within the service list already a long time ago.
There was only an anomaly in case the first service in a list was hidden. As this
anomaly was solved this patch is really not required anymore.