epgcache: Move savetofile routine from servicedvbrecord to epgcache
servicedvbrecord was the one and only user for two lookup functions
that required the cache to be locked from the outside. Move this
save-EIT-to-file routine to the epgcache, so that other objects no
longer need to know about the internal workings of the cache.

Remove the now unused lookupEvents methods that were recursively
locking the cache mutex.

This also removes lowlevel/eit.h from the header, only epgcache needs
to know about the EIT internals.