Fix Bug #301 - Better handling of HelpableActionMaps with entries with no help
If a HelpableActionMapwith a mapping with no help has a higher (lower number)
priority than another mapping with the same action name in another
HelpableActionMap, the help for the overridden mapping is shown in the help.

This fixes that problem, and implementing that fix required dealing with
helpableButtonSetupActionMap mappings, which I previously wasn't aware of.

This fix is required for implementation of UP/DOWN as switchable between
opening the channel list and initiating skipping.

Minor PEP8 fix in [ActionMap].