Prototype switchable up/down skips: switch UP/DOWN action without GUI restart
[InfoBarGenerics, Timeshift]
Change the mechanism for switching the UP/DOWN action from initialising
the ActionMaps differently to using ActionMap.setEnable().

This caused some problems with SeekActionsPTS actions overriding
TimeshiftActivateActions and TimeshiftActivateActionsUpDown after
the first time timeshift was used. The three ActionMaps had the
same priority, so I worked around the problem by making
TimeshiftActivateActions and TimeshiftActivateActionsUpDown higher
priority. However, the proble is really that SeekActionsPTS isn't
being disabled when timeshift exits.

Added space fill to times in skip help so that they sort in numerical
order.

[HelpMenuList]
Now merges ActionMap help entries that have the same description.
This was needed to merge the help for the skips for L/R with the
help for the skips for U/D.

Also tiny efficiency tweak in detecting whether an indented list
will be used.

[setup.xml]
Remove the "Requires restart" text from the "Enable skip on UP/UDOWN"
description text.