Prototype switchable up/down skips: auto-generate skip help in InfoBarSeekActionMap
[InfoBarGenerics, Timeshift, KeyBindings]
Refactored the skip time selection in InfoBarSeekActionMap so that
the skip time can be used to both implement the skips and
to generate the help strings for the skip buttons.

This simplified the controls for the UP/DOWN skips setting.

An additional useful side-effect is that all screens that use
InfoBarSeek get help strings for their skip buttons.