Prototype switchable up/down skips: Allow MEDIA key to return to movie selection list while playing movie.
[keymap.xml]
When UP/DOWN are configured for skips, there was no way to return
the movie selection list while playing a movie.

Added KEY_MEDIA to also perform this function. Not conditional on
the UP/DOWN skip setting, since KEY_MEDIAand KEY_FILE (not on T3
remote) also perform this function.