Prototype switchable up/down skips: Add percentage jumps to position to Seekpar and tidy up generated skip help
[Seekbar]
Number keys now jump to percentage positions in the recording/timeshift
buffer when the Seekbar is showing: 0 => start, 10 => 10% ... 9 => 90%.

[InfoBarSeek]
Fixed problem that the generated help text for seekdef: actions was
fixed at the time that the InfoBarSeekActionMap instabces are
created, and so they don't reflect any changes to the skip settings.

This fix also fixed the problem that the seekdef: help strings
weren't being flagged as user-configurable.

Also reverted the space-filling of the numbers in the skip help.
The padding made positive numbers sort correctly, but not the
negative ones.