Prototype switchable up/down skips: Stop channel selection list appearing on UP/DOWN shortly after channel change
[Timeshift]
If UP/DOWN are enabled as skips, rather than for entering channel
selection, if UP or DOWN is pressed during the delay before starting
timeshift, the channel list is shown, rather than no action being
taken.

Changed the condition on enabling the timeshift activation of UP/DOWN
so they are active from when the new channel is switched to, and
make activateTimeshiftEnd() safe to call if timeshiftEnabled() does
not (yet) return True.

Also removed the change of priority on TimeshiftActivateActions and
TimeshiftActivateActionsUpDown. The change seems to have been
unnecessary, and it may have actually been the symptoms of thhe
problem that the main pert of this fix changes.