Fix Bug #358 - Movie Player focus not reliably on first recording on entry
[MovieList]
Explicitly search for first recording to set MovieList.firstFileEntry,
instead of relying on the first recording being at index numberOfDirs.

The search is after all changes to MovieList.list in method load().

[MovieSelection]
Change the MENU>Settings>Sort "shuffle" entry text to "Shuffle" to
match capitalisation in the rest of the list.