Add IceTV timer icons
[RecordTimer]

Extend flagging returned by isInTimer() to indicate both Autotimers
and IceTV timers.

[EPGList]

Display IceTV as well as Autotimer icons for generated timers.

The EPGList code can display both an IceTV and an Autotimer icon for
an EPG entry, if there are two such timers for the entry, but
RecordTimer will only return the flagging for a single timer.

[TimerList]

Display IceTV as well as Autotimer icons for generated timers.