Fix Information>Network crash
Fixed crash when MENU>Setup>Information>Network is used when the
network status ping test fails. When the screen exits, the callback
for iNetwork.checkNetworkState() can still be called. The screen's
dict elements are deleted by the screen close, and a reference
through one of them in SystemNetworkInfo.checkNetworkCB() causes
the crash.

Now checks for the existence of the dict element to avoid crash.