Fix Bug #369 - Timeshift with up/down keys has long press issues.
The bug has two causes.

The first is that enabling the SeekActionsPTS ActionMap does not
happen synchronously with the
InfobarTimeshiftActivateActions/timeshiftActivateEnd action. It is
delayed until the next iPlayableService.evSeekableStatusChanged.

This means that incoming keys (and especially in this case repeats
of the UP and DOWN keys) can be processed with the old ActionMap
settings.

The second cause is that the InfobarChannelSelection/switchChannelUp
& switchChannelDown unnecessarily set their "r" repeat flag, so
that the next UP/DOWN prepeat is processed with the InfobarChannelSelection
ActionMap instead of SeekActionsPTS.

It's unclear whether the SeekActionsPTS ActionMap can be activated
synchronously, so that code has been left unaltered, and the
unnecessary "r" flags have been removed from
InfobarChannelSelection/switchChannelUp & switchChannelDown.

This commit also includes a fix to setup.xml to correct the description
text for the Skip time setting entries for the UP and DOWN skips.