Fix Bug #369 - Timeshift with up/down keys has long press issues.
Added a InfobarChannelSelectionUpDown context & ChannelSelectionUpDown
that are disabled when UP/DOWN are enabled for skips and timeshift
is activated.

The change of the InfobarChannelSelection UP/DOWN actions (now in
InfobarChannelSelectionUpDown) from having flags "mr" to having
just "m" that was made in the previous commit has been retained,
because having the repeat flag set on those keymap entries is
unnecessary.

Also properly split InfobarTimeshiftActivateUpDownActions from
InfobarTimeshiftActivateActions to avoid unnecessary warnings in
the debug log.