Fix Bug #370: Incorrect "save settings" warning when exiting Movie Player config
ConfigNumber is passed an integer as its default. It then passes
that directly to ConfigText as its default and that is saved in
ConfigText as an integer, when it should be a string. When
ConfigNumber.isChanged() compares the current value with the default,
it compares the string current value with the integer default, and
they always compare unequal, even when they represent the same
number.

That causes the "Disk space to reserve" item in Menu Player's
Settings screen to always indicate that it has been changed.

Convert the default parameter to ConfigNumber to a string when it
is passed to ConfigText.