Fix #367: Fix sort order in Timeshift save recording (select event) popup
Filter file list for popup on creation so that only the required
file namess are presented for sorting and insertion into the popup
list.

Sort numerical descending on the PTS file name's number suffix
instead of lexicographically ascending on the whole filename. This
will sort most-recent-first.

Shorten title of popup so that it doesn't get cut off by the edge
of the popup's ChoiceBox.

Use strip() to remove extraneous whitespace (including line terminator)
from metafile data lines, instead of relying on [0:-1] to simply
cut off the last character.