Fix Bug (partial) #388: Filenames containing some shell metacharacters are not handled correctly in some Tasks or Console commands
Fixes problems with shell metacharacters in filenames in uesr
operations Move, Delete (move to Trash) and Copy in [MovieSelection],
though the fix is in [CopyFiles].

This fix affects the above operations and any operations that use
Tools.Directories.moveFiles(). However, I haven't been able to find
any references to that function in the enigma2 source, the
enigma2-plugins source or branding-module source.