Fix Bug (partial) #388: Filenames containing some shell metacharacters are not handled correctly [FileTransfer, FileCommander]
Added function Tools.Directories.shellquote() to safely quote
strings that will be passed to the shell for execution.

In [FileTransfer] replace the use of double quotes for shell quoting
with shellquote(). This is *not* strictly backwards-compatible, but
the intenton of the code appears to be to quote the filenames passed to
FileTransferTask; this simply provides a correct implementation.

The use of FileTransferJob/FileTransferTask seems quite restricted.
Out of enigma2, enigma2-plugins, branding-module the classes are only
used in [FileCommander].

Minor correction to debug print in [FileCommander].