Fix Bug (partial) #388: Filenames containing some shell metacharacters are not handled correctly - use of eConsoleApp & Popen [FileCommander, FileCommander.addons]
[FileCommander.plugin]
[FileCommander.addons.ipk]
[FileCommander.addons.tar]
[FileCommander.addons.unrar]
[FileCommander.addons.unzip]

Replace the use of shell-interpreted strings in eConsoleApp & Popen
for operations on archives and packages in all modules.

Replaced the fixed unpack destination strings "/media/hdd/movie/"
in the addons with a reference to config.usage.default_path.

Simplified the processing of the output of the subprocess.Popen()
calls.

[FileCommander.addons.ipk]
Changed the use of 'ar' in operations on .ipk files to use tar (.ipk
files are no longer ar format).  Remove unpack options for .ipk
files; only "Show contents" and "Install" are now supported.

"Show contents" could not be done using execvp()-based commands
because of the risk of deadlock reading the outputs. Uses a shell
pipeline with uncontrolled strings quoted by
Tools.Directories.shellquote().

[FileCommander.addons.unzip]
Changed the font for the display of the contents of the ZIP file to a
fixed-width font.

General:
Small amount of tidying-up of UI strings, changed some comments, debug prints and
local variable named from German to English.