Fix Bug (partial) #388: Filenames containing some shell metacharacters are not handled correctly - tidied up use of filename extensions [FileCommander.FileListmod, FileCommander.addons.key_actions]
These changes affect the display of file lists in the FileCommander.

[FileCommander.FileListmod]
[FileCommander.addons.key_actions]
Added "tgz" as a recognised tar file.

Distinguished (in the code, though not in the icons) between
a .tar.tgz file, which tar can process, and other .gz files
which tar cannot process.

Simplified and made the extraction of extensinos more robust
by using os.patc.splitext() instead of DIY user code.

Where possible, made the tests for actions that match extensions
to use "filetype in MOVIE_EXTENSIONS" instead of
"testFileName.endswith(tuple(MOVIE_EXTENSIONS))".
The *_EXTENSIONS objects imported from Components.MovieList
are frozensets.

[FileCommander.FileListmod]
Removed the regular expression "r\d+$" from being an key
in EXTENSIONS.

Moved all code mapping filenames to icons from inline to
a function. This function also matches the extension on
"^r\d+$" to be recognised as a RAR file (the match does need
to be anchored at both ends).

[FileCommander.addons.key_actions]
Changed the "\\n" in the heading of the "What do you want to play?"
ChoiceBox so that it formats correctly.

Added a TEXT_EXTENSIONS frozenset to match text file extensions instead
of a string of "testFileName.endswith(".txt") or ..." tests.

[FileCommander.addons.type_utils]
[FileCommander.addons.key_actions]
Remove unused references to eConsoleAppContainer