socket: use getaddrinfo
gethostbyname is obsolete, and getaddrinfo allows
us to avoid IPv4/IPv6 dependencies.
Also, getaddrinfo can return multiple addresses,
allowing multihomed and/or multiprotocol services
to be bound or connected.

Same as d19775591b88bdc78689fd54798fdd490ba07f21
(which was reverted because of an uninitialized
eSocket::socketdesc member, fixed in this commit)