Refactor archiver unpack code [FileCommander.addons]
Moved code common to ipk, tar, unrar and unzip addons into new
unarchiver module, removed unnecessary duplication of code in the
*.unpackModus() methods and used new unarchiver.unpackPopen() and
unpackEConsoleApp() to run common command execution code.

Used Components.MultiContent functions to build eListboxPythonMultiContent
templates.

Used grey as the text colour in the progress bars instead of white
so that the text is clearer once the progress bar goes past 50%.

In unrar, use resolveFilename(SCOPE_CURRENT_PLUGIN, ...) to build
the pathname of the unrar command embedded in the package instead
of an absolute pathname.

Removed lots of unused imports.