Fix #363: Auto fade of infobar on pause when playing media files
Fade the infobar on pause.

The infobar times out on pause only if Behavior of PAUSE when paused is set to Play. In other settings its behaviour remains unchanged.

OK shows the infobar when paused and Behavior of PAUSE when paused is set to Play. In other settings, OK retains its previous role of resuming normal play from PAUSE. This is because if the setting is Play, PAUSE/PLAY resumes normal play. In other settings it does not, and OK is the only way back. This is also the main reason why I only fade the infobar in the Play mode. In the other mode there's no simple way to just get the inobar back.

I have tried to clarify the information (particularly the description) for the Behavior of PAUSE when paused setting (it's now also called Behavior of PAUSE/PLAY when paused because it actually has no effect on the PAUSE button on remotes that have separate PLAY and PAUSE).